requirements.txt

numpy
pandas
scikit-learn
pydantic
fastapi
uvicorn
PyYAML
streamlit
joblib
